fix: move _run_async to separate module to avoid circular import and recursion
The previous fix put _run_async in tvdb_matcher.py, then replace_all
turned asyncio.run() calls inside the function itself into recursive
_run_async() calls. Moved to _async_compat.py as a standalone module
with no circular dependencies.
