+++ arm/services/matching/tvdb_matcher.py
@@ -11,45 +11,15 @@ against TVDB episode runtimes.  Supports:
 
 from __future__ import annotations
 
-import asyncio
 import logging
-from concurrent.futures import Future
-from threading import Thread
 from typing import Any
 
 import arm.config.config as cfg
+from arm.services.matching._async_compat import run_async as _run_async
 from arm.services.matching.base import MatchResult, MatchStrategy, TrackMatch
 from arm.services.matching.cross_disc import get_excluded_episodes
 
 log = logging.getLogger(__name__)
-
-
-def _run_async(coro):
-    """Run an async coroutine from sync code, regardless of event loop state.
-
-    When called from a context with no running event loop (e.g. the ripper
-    process), uses ``_run_async()``.  When called from inside an existing
-    loop (e.g. a FastAPI endpoint), runs the coroutine in a background
-    thread to avoid the "cannot be called from a running event loop" error.
-    """
-    try:
-        asyncio.get_running_loop()
-    except RuntimeError:
-        # No running loop — safe to use _run_async()
-        return _run_async(coro)
-
-    # Already inside an event loop — run in a new thread with its own loop
-    result_future: Future = Future()
-
-    def _thread_target():
-        try:
-            result_future.set_result(_run_async(coro))
-        except Exception as e:
-            result_future.set_exception(e)
-
-    t = Thread(target=_thread_target, daemon=True)
-    t.start()
-    return result_future.result(timeout=60)
 
 
 class TvdbMatcher(MatchStrategy):
